feat(protocols): наружу можно выйти через SOCKS5 и HTTP-прокси
Протокол был один, и всё вокруг было устроено так, будто других не бывает:
форма сервера знала поля Hysteria 2 наизусть, а «Добавить сервер» сразу её и
открывало.

Протоколы. SOCKS5 (RFC 1928, 1929) и HTTP CONNECT — в открытую и под TLS. Ни у
того, ни у другого своего шифрования нет, и нужны они для выхода через свой
прокси: туннель `ssh -D`, соседний контейнер, корпоративная сеть. Разрешение
имени и перебор адресов вынесены в `proto::connect`: они одинаковы у всех, а
копия в каждом протоколе — верный способ получить разные ответы на вопрос, что
делать, когда у сервера два адреса.

Окно. Сначала выбор протокола, потом форма: поля формы и есть протокол, и
список внутри неё менял бы половину экрана прямо под руками. Схему полей окно
держит у себя (`forms/protocol`), а не спрашивает протокол: `protocols/*`
подключает один `engine`, и стрелка отсюда туда означала бы, что каждый новый
протокол правит заодно и интерфейс. Ловит окно только то, что видно без сети, —
пустое обязательное поле и опечатку в адресе; настоящую проверку делает сам
протокол.

Выбор файла для правил. Список запущенного отвечает не на все вопросы: правило
пишут и для игры, в которую сегодня не играли, и для обновлятора, который
просыпается сам. Окно берётся системное — своё пришлось бы учить сетевым
дискам, ярлыкам и правам на чужие каталоги. Приведение пути к сравнимому виду
стало одно на всех (`core::path`): путь приходит с трёх сторон, и копия на
каждой однажды разошлась бы молча.

Заодно вычищены предупреждения clippy: `Default` через derive, `as_chunks`
вместо `chunks_exact(2)`, `sort_by_key`, снятый общий постфикс у
`SettingsMessage`. И семь файлов вернулись с CRLF на LF — остальной репозиторий
весь на LF.

+++ crates/platform/src/bind.rs
@@ -143,7 +143,7 @@ fn set_option(
         (IPPROTO_IP.0, IP_UNICAST_IF)
     };
 
-    let code = unsafe { setsockopt(SOCKET(socket as usize), level, name as i32, Some(&bytes)) };
+    let code = unsafe { setsockopt(SOCKET(socket as usize), level, name, Some(&bytes)) };
     check(code)
 }
